Refactor test helper functions
test helper functions are shifted to a new module testutils.rs
Required modification made in other modules
